Fix potential out-of-bound write when initializing absinfo structures
An out-of-bounds access could have happen when writing a device's absinfo. If the
devinfo's abs array contained more bits then there are elements in the absinfo
array (because of rounding), the absinfo's copy operation operated on invalid memory.

This problem has been fixed by only copying absinfo fields that have their bit set
in the abs array. This should never be the case for any extra bits.